Move class reference updating to command
